♻️ refactor(aiAgent): extract InterventionController and threadRunHooks
Move interruption/approval lifecycle methods into a dedicated
InterventionController collaborator (constructor-injected deps) and the
four thread hook factories into hooks/threadRunHooks.ts as free
functions taking their model deps explicitly.

AiAgentService keeps thin delegating methods, so the public surface is
unchanged. Pure code motion — method bodies unchanged apart from
this.X → deps.X.

LOBE-13708
